requirements.txt
Flask==3.1.2
joblib==1.5.2
pandas==2.3.2
scikit-learn==1.7.2
xgboost==3.0.5
flask-cors==6.0.1
python-dotenv==1.1.1
openai==1.109.1